refactor: create ui package
