+++ src/main/java/commands/AddCommand.java
@@ -7,7 +7,7 @@ import task.Event;
 import task.Task;
 import task.TaskList;
 import task.ToDo;
-import app.Ui;
+import ui.Ui;
 import java.time.format.DateTimeParseException;
 
 /**
