+++ src/main/java/commands/MarkCommand.java
@@ -4,7 +4,7 @@ import storage.Storage;
 import storage.StorageException;
 import task.Task;
 import task.TaskList;
-import app.Ui;
+import ui.Ui;
 
 /**
  * Represents a command to mark a task as done.
