+++ src/main/java/commands/UnmarkCommand.java
@@ -4,7 +4,7 @@ import storage.Storage;
 import storage.StorageException;
 import task.Task;
 import task.TaskList;
-import app.Ui;
+import ui.Ui;
 
 /**
  * Represents a command to unmark a task (mark as not done).
