+++ src/main/java/logic/ChatBot.java
@@ -6,7 +6,7 @@ import storage.Storage;
 import storage.StorageException;
 import task.Task;
 import task.TaskList;
-import app.Ui;
+import ui.Ui;
 
 /**
  * Represents a Chatbot that manages tasks using the Command pattern.
