+++ src/main/java/ui/Ui.java
@@ -1,4 +1,4 @@
-package app;
+package ui;
 
 import task.Task;
 import java.util.Scanner;
